refactor(model): 패키지 분리
