+++ src/main/java/blackjack/model/Hands.java
@@ -1,5 +1,6 @@
 package blackjack.model;
 
+import blackjack.model.card.Card;
 import java.util.ArrayList;
 import java.util.List;
