+++ src/main/java/blackjack/model/cardDeck/PickStrategy.java
@@ -1,5 +1,6 @@
-package blackjack.model;
+package blackjack.model.cardDeck;
 
+import blackjack.model.card.Card;
 import java.util.List;
 
 public interface PickStrategy {
