+++ src/main/java/blackjack/model/cardDeck/RandomPickStrategy.java
@@ -1,5 +1,6 @@
-package blackjack.model;
+package blackjack.model.cardDeck;
 
+import blackjack.model.card.Card;
 import java.util.List;
 import java.util.Random;
